테스트 코드 작성 및 기타 soft delete 관련 코드 에러 수정

+++ src/main/java/com/softwarecampus/backend/service/board/BoardServiceImpl.java
@@ -7,12 +7,12 @@ import com.softwarecampus.backend.exception.board.BoardErrorCode;
 import com.softwarecampus.backend.exception.board.BoardException;
 import com.softwarecampus.backend.repository.board.*;
 import com.softwarecampus.backend.repository.user.AccountRepository;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -30,6 +30,7 @@ public class BoardServiceImpl implements BoardService {
     private final BoardRecommendRepository boardRecommendRepository;
     private final CommentRecommendRepository commentRecommendRepository;
 
+    @Transactional(readOnly = true)
     @Override
     public Page<BoardListResponseDTO> getBoards(int pageNo, BoardCategory category, String searchType, String searchText) {
         PageRequest pageRequest = PageRequest.of(pageNo - 1, 10, Sort.by("id").descending());
@@ -38,11 +39,11 @@ public class BoardServiceImpl implements BoardService {
             return boardRepository.findBoardsByCategory(category, pageRequest);
         } else {
             if ("title".equals(searchType)) {
-                return boardRepository.findBoardsByTitle(category, searchText, pageRequest);
+                return boardRepository.findBoardsByTitle(category, "%" + searchText + "%", pageRequest);
             } else if ("text".equals(searchType)) {
-                return boardRepository.findBoardsByText(category, searchText, pageRequest);
+                return boardRepository.findBoardsByText(category, "%" + searchText + "%", pageRequest);
             } else if ("title+text".equals(searchType)) {
-                return boardRepository.findBoardsByTitleAndText(category, searchText, pageRequest);
+                return boardRepository.findBoardsByTitleAndText(category, "%" + searchText + "%", pageRequest);
             } else {
                 throw new BoardException(BoardErrorCode.SEARCHTYPE_MISSMATCH);
             }
@@ -50,9 +51,14 @@ public class BoardServiceImpl implements BoardService {
     }
 
     //
+    @Transactional(readOnly = true)
     @Override
     public BoardResponseDTO getBoardById(Long id, Long userId) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new BoardException(BoardErrorCode.BOARD_NOT_FOUND));
+        //삭제된 글 조회 불가 처리
+        if (!board.isActive()) {
+            throw new BoardException(BoardErrorCode.BOARD_NOT_FOUND);
+        }
         board.setHits(board.getHits() + 1);
         BoardResponseDTO boardResponseDTO = BoardResponseDTO.from(board);
         if (userId != null) {
@@ -91,6 +97,10 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public void updateBoard(BoardUpdateRequestDTO boardUpdateRequestDTO, MultipartFile[] files) {
         Board board = boardRepository.findById(boardUpdateRequestDTO.getId()).orElseThrow(() -> new BoardException(BoardErrorCode.BOARD_NOT_FOUND));
+        //삭제된 글 수정 불가 처리
+        if (!board.isActive()) {
+            throw new BoardException(BoardErrorCode.BOARD_NOT_FOUND);
+        }
         List<BoardAttach> boardAttachList = board.getBoardAttaches();
         if (files != null && files.length > 0) {
             if (boardAttachList.size() > 0) {
@@ -114,6 +124,10 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public void deleteBoardById(Long id) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new BoardException(BoardErrorCode.BOARD_NOT_FOUND));
+        //삭제된 글 다시삭제 불가 처리
+        if (!board.isActive()) {
+            throw new BoardException(BoardErrorCode.BOARD_NOT_FOUND);
+        }
         board.markDeleted();
     }
 
@@ -123,7 +137,10 @@ public class BoardServiceImpl implements BoardService {
 
         Account account = accountRepository.findById(userId).get();
         Board board = boardRepository.findById(boardId).orElseThrow(() -> new BoardException(BoardErrorCode.BOARD_NOT_FOUND));
-
+        //삭제된 글 조회 불가 처리
+        if (!board.isActive()) {
+            throw new BoardException(BoardErrorCode.BOARD_NOT_FOUND);
+        }
         if (boardRecommendRepository.findByBoardIdAndUserId(boardId, userId) != null) {
             throw new BoardException(BoardErrorCode.ALREADY_RECOMMEND_BOARD);
         }
@@ -139,6 +156,9 @@ public class BoardServiceImpl implements BoardService {
     public void unRecommendBoard(Long boardId, Long userId) {
 
         BoardRecommend boardRecommend = boardRecommendRepository.findByBoardIdAndUserId(boardId, userId);
+        if (boardRecommend != null && !boardRecommend.getBoard().isActive()) {
+            throw new BoardException(BoardErrorCode.BOARD_NOT_FOUND);
+        }
         if (boardRecommend == null) {
             throw new BoardException(BoardErrorCode.NOT_RECOMMEND_BOARD);
         }
@@ -150,10 +170,21 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public Long createComment(CommentCreateRequestDTO commentCreateRequestDTO, Long userId) {
         Board board = boardRepository.findById(commentCreateRequestDTO.getBoardId()).orElseThrow(() -> new BoardException(BoardErrorCode.BOARD_NOT_FOUND));
+        //삭제된 글 조회 불가(댓글달기 불가)
+        if (!board.isActive()) {
+            throw new BoardException(BoardErrorCode.BOARD_NOT_FOUND);
+        }
         //실제론 1L 대신 userId가 인자로 전달
         Account account = accountRepository.findById(1L).get();
         Comment comment = commentCreateRequestDTO.toEntity(board, account);
-
+        Comment topComment = null;
+        if (commentCreateRequestDTO.getTopCommentId() != null) {
+            topComment = commentRepository.findById(commentCreateRequestDTO.getTopCommentId()).orElseThrow(() -> new BoardException(BoardErrorCode.COMMENT_NOT_FOUND));
+        }
+        if (!topComment.isActive()) {
+            throw new BoardException(BoardErrorCode.COMMENT_NOT_FOUND);
+        }
+        comment.setTopComment(topComment);
         commentRepository.save(comment);
 
         return comment.getId();
@@ -163,6 +194,10 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public void updateComment(CommentUpdateRequestDTO commentUpdateRequestDTO) {
         Comment comment = commentRepository.findById(commentUpdateRequestDTO.getId()).orElseThrow(() -> new BoardException(BoardErrorCode.COMMENT_NOT_FOUND));
+        //삭제된 댓글 수정 불가 처리
+        if (!comment.isActive()) {
+            throw new BoardException(BoardErrorCode.COMMENT_NOT_FOUND);
+        }
         commentUpdateRequestDTO.updateEntity(comment);
     }
 
@@ -170,6 +205,10 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public void deleteComment(Long id) {
         Comment comment = commentRepository.findById(id).orElseThrow(() -> new BoardException(BoardErrorCode.COMMENT_NOT_FOUND));
+        //삭제된 댓글 재삭제 불가 처리
+        if (!comment.isActive()) {
+            throw new BoardException(BoardErrorCode.COMMENT_NOT_FOUND);
+        }
         comment.markDeleted();
     }
 
@@ -178,6 +217,10 @@ public class BoardServiceImpl implements BoardService {
     public void recommendComment(Long commentId, Long userId) {
         Account account = accountRepository.findById(userId).get();
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new BoardException(BoardErrorCode.COMMENT_NOT_FOUND));
+        //삭제된 댓글 추천 불가 처리
+        if (!comment.isActive()) {
+            throw new BoardException(BoardErrorCode.COMMENT_NOT_FOUND);
+        }
         if (commentRecommendRepository.findByCommentIdAndUserId(commentId, userId) != null) {
             throw new BoardException(BoardErrorCode.ALREADY_RECOMMEND_COMMENT);
         }
@@ -192,9 +235,14 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public void unRecommendComment(Long commentId, Long userId) {
         CommentRecommend commentRecommend = commentRecommendRepository.findByCommentIdAndUserId(commentId, userId);
+        if (commentRecommend != null && !commentRecommend.getComment().isActive()) {
+            throw new BoardException(BoardErrorCode.COMMENT_NOT_FOUND);
+        }
         if (commentRecommend == null) {
             throw new BoardException(BoardErrorCode.NOT_RECOMMEND_COMMENT);
         }
+
+
         commentRecommendRepository.delete(commentRecommend);
     }
